@@ -1,5 +1,5 @@
 {
-  "name": "troydavisson/phrets",
+  "name": "qoboltd/phrets",
   "type": "library",
   "license": "MIT",
   "description": "RETS library in PHP",
@@ -7,27 +7,41 @@
     "rets",
     "mls"
   ],
-  "homepage": "http://github.com/troydavisson/PHRETS",
+  "homepage": "https://github.com/QoboLtd/PHRETS",
   "authors": [
+    {
+      "name": "Qobo Ltd",
+      "email": "hello@qobogroup.com",
+      "homepage": "https://qobogroup.com/",
+      "role": "Developer"
+    },
     {
       "name": "Troy Davisson",
       "email": "troy.davisson@gmail.com",
       "homepage": "http://troda.com",
       "role": "Developer"
+    },
+    {
+      "name": "Max Lispky",
+      "email": "maxlipsky@gmail.com",
+      "role": "Developer"
     }
   ],
   "require": {
-    "php": ">=5.5.0",
-    "guzzlehttp/guzzle": ">=6.0",
-    "illuminate/container": ">=4.2.0",
-    "illuminate/support": ">=4.2.0",
-    "league/csv": ">=6.0"
+    "php": ">=8.1",
+    "guzzlehttp/guzzle": ">=7.2",
+    "illuminate/container": ">=9.0",
+    "illuminate/support": ">=9.0",
+    "league/csv": ">=9.7",
+    "ext-simplexml": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "5.*",
-    "psr/log": "1.*",
-    "monolog/monolog": "1.10.*",
-    "gsaulmon/guzzlerecorder": "2.*"
+    "phpunit/phpunit": "9.5.*",
+    "psr/log": "3.*",
+    "monolog/monolog": "3.1.*",
+    "rector/rector": "^0.13.8",
+    "phpstan/phpstan": "^1.8",
+    "friendsofphp/php-cs-fixer": "^3.8"
   },
   "autoload": {
     "psr-4": {
