@@ -1,100 +1,75 @@
-<?php namespace PHRETS\Models;
+<?php
+
+namespace PHRETS\Models;
 
 use Illuminate\Support\Arr;
 
-class Bulletin
+class Bulletin implements \Stringable
 {
-    protected $body = null;
-    protected $details = [];
+    protected ?string $body = null;
+    protected array $details = [];
 
-    /**
-     * @param array $details
-     */
-    public function __construct($details = [])
+    public function __construct(array $details = [])
     {
-        if ($details and is_array($details)) {
+        if ($details && is_array($details)) {
             $this->details = array_change_key_case($details, CASE_UPPER);
         }
     }
 
-    /**
-     * @return mixed
-     */
-    public function getBody()
+    public function getBody(): ?string
     {
         return $this->body;
     }
 
     /**
-     * @param mixed $body
      * @return $this
      */
-    public function setBody($body)
+    public function setBody($body): static
     {
         $this->body = $body;
+
         return $this;
     }
 
-    public function setDetail($name, $value)
+    public function setDetail(string $name, $value): static
     {
         $this->details[strtoupper($name)] = $value;
+
         return $this;
     }
 
-    /**
-     * @param $name
-     * @return mixed
-     */
-    public function getDetail($name)
+    public function getDetail(string $name): mixed
     {
         return Arr::get($this->details, strtoupper($name));
     }
 
-    /**
-     * @return mixed
-     */
     public function getMemberName()
     {
         return $this->getDetail('MemberName');
     }
 
-    /**
-     * @return mixed
-     */
     public function getUser()
     {
         return $this->getDetail('User');
     }
 
-    /**
-     * @return mixed
-     */
     public function getBroker()
     {
         return $this->getDetail('Broker');
     }
 
-    /**
-     * @return mixed
-     */
     public function getMetadataVersion()
     {
         return $this->getDetail('MetadataVersion');
     }
 
-    /**
-     * @return mixed
-     */
     public function getMetadataTimestamp()
     {
         return $this->getDetail('MetadataTimestamp');
     }
 
-    /**
-     * @return string
-     */
-    public function __toString()
+    public function __toString(): string
     {
-        return (string)$this->body;
+        return (string) $this->body;
     }
 }
