@@ -1,4 +1,6 @@
-<?php namespace PHRETS\Parsers\GetObject;
+<?php
+
+namespace PHRETS\Parsers\GetObject;
 
 use PHRETS\Http\Response;
 use PHRETS\Models\BaseObject;
@@ -6,9 +8,9 @@ use PHRETS\Models\RETSError;
 
 class Single
 {
-    public function parse(Response $response)
+    public function parse(Response $response): BaseObject
     {
-        $obj = new BaseObject;
+        $obj = new BaseObject();
         $obj->setContent(($response->getBody()) ? $response->getBody()->__toString() : null);
         $obj->setContentDescription($response->getHeader('Content-Description'));
         $obj->setContentSubDescription($response->getHeader('Content-Sub-Description'));
@@ -21,33 +23,33 @@ class Single
 
         if ($this->isError($response)) {
             $xml = $response->xml();
-            
-            $error = new RETSError;
-            
+
+            $error = new RETSError();
+
             if (isset($xml['ReplyCode'])) {
                 $error->setCode((string) $xml['ReplyCode']);
             }
             if (isset($xml['ReplyText'])) {
                 $error->setMessage((string) $xml['ReplyText']);
             }
-            
+
             $obj->setError($error);
         }
 
         return $obj;
     }
 
-    protected function isError(Response $response)
+    protected function isError(Response $response): bool
     {
         if ($response->getHeader('RETS-Error') == 1) {
             return true;
         }
 
-        $content_type = $response->getHeader('Content-Type');
-        if ($content_type and strpos($content_type, 'text/xml') !== false) {
+        $content_type = (string) $response->getHeader('Content-Type');
+        if ($content_type && str_contains($content_type, 'text/xml')) {
             $xml = $response->xml();
 
-            if (isset($xml['ReplyCode']) and $xml['ReplyCode'] != 0) {
+            if (isset($xml['ReplyCode']) && $xml['ReplyCode'] != 0) {
                 return true;
             }
         }
