@@ -1,4 +1,6 @@
-<?php namespace PHRETS;
+<?php
+
+namespace PHRETS;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\Cookie\CookieJar;
@@ -20,8 +22,6 @@ use Psr\Http\Message\ResponseInterface;
 
 class Session
 {
-    /** @var Configuration */
-    protected $configuration;
     /** @var Capabilities */
     protected $capabilities;
     /** @var Client */
@@ -34,32 +34,29 @@ class Session
     /** @var ResponseInterface */
     protected $last_response;
 
-    public function __construct(Configuration $configuration)
+    public function __construct(protected Configuration $configuration)
     {
-        // save the configuration along with this session
-        $this->configuration = $configuration;
-
         $defaults = [];
 
         // start up our Guzzle HTTP client
         $this->client = PHRETSClient::make($defaults);
 
-        $this->cookie_jar = new CookieJar;
+        $this->cookie_jar = new CookieJar();
 
         // start up the Capabilities tracker and add Login as the first one
-        $this->capabilities = new Capabilities;
+        $this->capabilities = new Capabilities();
         $this->capabilities->add('Login', $configuration->getLoginUrl());
     }
 
     /**
-     * PSR-3 compatible logger can be attached here
+     * PSR-3 compatible logger can be attached here.
      *
      * @param $logger
      */
     public function setLogger($logger)
     {
         $this->logger = $logger;
-        $this->debug("Loading " . get_class($logger) . " logger");
+        $this->debug('Loading ' . $logger::class . ' logger');
     }
 
     /**
@@ -69,14 +66,14 @@ class Session
      */
     public function Login()
     {
-        if (!$this->configuration or !$this->configuration->valid()) {
-            throw new MissingConfiguration("Cannot issue Login without a valid configuration loaded");
+        if (!$this->configuration || !$this->configuration->valid()) {
+            throw new MissingConfiguration('Cannot issue Login without a valid configuration loaded');
         }
 
         $response = $this->request('Login');
 
         $parser = $this->grab(Strategy::PARSER_LOGIN);
-        $xml = new \SimpleXMLElement((string)$response->getBody());
+        $xml = new \SimpleXMLElement((string) $response->getBody());
         $parser->parse($xml->{'RETS-RESPONSE'}->__toString());
 
         foreach ($parser->getCapabilities() as $k => $v) {
@@ -87,6 +84,7 @@ class Session
         if ($this->capabilities->get('Action')) {
             $response = $this->request('Action');
             $bulletin->setBody($response->getBody()->__toString());
+
             return $bulletin;
         } else {
             return $bulletin;
@@ -98,11 +96,13 @@ class Session
      * @param $type
      * @param $content_id
      * @param int $location
+     *
      * @return \PHRETS\Models\BaseObject
      */
     public function GetPreferredObject($resource, $type, $content_id, $location = 0)
     {
         $collection = $this->GetObject($resource, $type, $content_id, '0', $location);
+
         return $collection->first();
     }
 
@@ -112,10 +112,12 @@ class Session
      * @param $content_ids
      * @param string $object_ids
      * @param int $location
+     *
      * @return Collection|BaseObject[]
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
-    public function GetObject($resource, $type, $content_ids, $object_ids = '*', $location = 0)
+    public function GetObject($resource, $type, $content_ids, $object_ids = '*', $location = 0): Collection|array
     {
         $request_id = GetObject::ids($content_ids, $object_ids);
 
@@ -127,7 +129,7 @@ class Session
                     'Type' => $type,
                     'ID' => implode(',', $request_id),
                     'Location' => $location,
-                ]
+                ],
             ]
         );
 
@@ -135,7 +137,7 @@ class Session
             $parser = $this->grab(Strategy::PARSER_OBJECT_MULTIPLE);
             $collection = $parser->parse($response);
         } else {
-            $collection = new Collection;
+            $collection = new Collection();
             $parser = $this->grab(Strategy::PARSER_OBJECT_SINGLE);
             $object = $parser->parse($response);
             $collection->push($object);
@@ -146,6 +148,7 @@ class Session
 
     /**
      * @return Models\Metadata\System
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
     public function GetSystemMetadata()
@@ -155,11 +158,11 @@ class Session
 
     /**
      * @param string $resource_id
+     *
      * @throws Exceptions\CapabilityUnavailable
      * @throws Exceptions\MetadataNotFound
-     * @return Collection|\PHRETS\Models\Metadata\Resource
      */
-    public function GetResourcesMetadata($resource_id = null)
+    public function GetResourcesMetadata($resource_id = null): Collection|Models\Metadata\Resource
     {
         $result = $this->MakeMetadataRequest('METADATA-RESOURCE', 0, 'metadata.resource');
 
@@ -178,7 +181,7 @@ class Session
 
     /**
      * @param $resource_id
-     * @return mixed
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
     public function GetClassesMetadata($resource_id)
@@ -190,17 +193,19 @@ class Session
      * @param $resource_id
      * @param $class_id
      * @param string $keyed_by
+     *
      * @return \Illuminate\Support\Collection|\PHRETS\Models\Metadata\Table[]
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
-    public function GetTableMetadata($resource_id, $class_id, $keyed_by = 'SystemName')
+    public function GetTableMetadata($resource_id, $class_id, $keyed_by = 'SystemName'): Collection|array
     {
         return $this->MakeMetadataRequest('METADATA-TABLE', $resource_id . ':' . $class_id, 'metadata.table', $keyed_by);
     }
 
     /**
      * @param $resource_id
-     * @return mixed
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
     public function GetObjectMetadata($resource_id)
@@ -211,7 +216,7 @@ class Session
     /**
      * @param $resource_id
      * @param $lookup_name
-     * @return mixed
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
     public function GetLookupValues($resource_id, $lookup_name)
@@ -224,8 +229,8 @@ class Session
      * @param $id
      * @param $parser
      * @param null $keyed_by
+     *
      * @throws Exceptions\CapabilityUnavailable
-     * @return mixed
      */
     protected function MakeMetadataRequest($type, $id, $parser, $keyed_by = null)
     {
@@ -236,11 +241,12 @@ class Session
                     'Type' => $type,
                     'ID' => $id,
                     'Format' => 'STANDARD-XML',
-                ]
+                ],
             ]
         );
 
         $parser = $this->grab('parser.' . $parser);
+
         return $parser->parse($this, $response, $keyed_by);
     }
 
@@ -249,10 +255,12 @@ class Session
      * @param $class_id
      * @param $dmql_query
      * @param array $optional_parameters
+     *
      * @return \PHRETS\Models\Search\Results
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
-    public function Search($resource_id, $class_id, $dmql_query, $optional_parameters = [], $recursive = false)
+    public function Search($resource_id, $class_id, $dmql_query, $optional_parameters = [], $recursive = false): Models\Search\Results
     {
         $dmql_query = Search::dmql($dmql_query);
 
@@ -263,21 +271,21 @@ class Session
             'QueryType' => 'DMQL2',
             'Count' => 1,
             'Format' => 'COMPACT-DECODED',
-            'Limit' => 99999999,
+            'Limit' => 99_999_999,
             'StandardNames' => 0,
         ];
 
         $parameters = array_merge($defaults, $optional_parameters);
 
         // if the Select parameter given is an array, format it as it needs to be
-        if (array_key_exists('Select', $parameters) and is_array($parameters['Select'])) {
+        if (array_key_exists('Select', $parameters) && is_array($parameters['Select'])) {
             $parameters['Select'] = implode(',', $parameters['Select']);
         }
 
         $response = $this->request(
             'Search',
             [
-                'query' => $parameters
+                'query' => $parameters,
             ]
         );
 
@@ -286,21 +294,25 @@ class Session
         } else {
             $parser = $this->grab(Strategy::PARSER_SEARCH);
         }
+
         return $parser->parse($this, $response, $parameters);
     }
 
     /**
      * @return bool
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
     public function Logout()
     {
         $this->request('Logout');
+
         return true;
     }
 
     /**
      * @return bool
+     *
      * @throws Exceptions\CapabilityUnavailable
      */
     public function Disconnect()
@@ -312,12 +324,15 @@ class Session
      * @param $capability
      * @param array $options
      * @param bool $is_retry
+     *
      * @return ResponseInterface
+     *
      * @throws CapabilityUnavailable
      * @throws RETSException
      */
     protected function request($capability, $options = [], $is_retry = false)
     {
+        $response = null;
         $url = $this->capabilities->get($capability);
 
         if (!$url) {
@@ -336,10 +351,10 @@ class Session
         $this->last_request_url = $url;
 
         try {
-            /** @var ResponseInterface $response */
+            /* @var ResponseInterface $response */
             if ($this->configuration->readOption('use_post_method')) {
                 $this->debug('Using POST method per use_post_method option');
-                $query = (array_key_exists('query', $options)) ? $options['query'] : null;
+                $query = $options['query'] ?? null;
 
                 // do not send query options in url, only in form_params
                 $local_options = $options;
@@ -353,7 +368,7 @@ class Session
                 $response = $this->client->request('GET', $url, $options);
             }
         } catch (ClientException $e) {
-            $this->debug("ClientException: " . $e->getCode() . ": " . $e->getMessage());
+            $this->debug('ClientException: ' . $e->getCode() . ': ' . $e->getMessage());
 
             if ($e->getCode() != 401) {
                 // not an Unauthorized error, so bail
@@ -377,8 +392,8 @@ class Session
                 throw $e;
             }
 
-            $this->debug("401 Unauthorized exception returned");
-            $this->debug("Logging in again and retrying request");
+            $this->debug('401 Unauthorized exception returned');
+            $this->debug('Logging in again and retrying request');
             // see if logging in again and retrying the request works
             $this->Login();
 
@@ -391,38 +406,36 @@ class Session
 
         if ($response->getHeader('Set-Cookie')) {
             $cookie = $response->getHeader('Set-Cookie');
-            if ($cookie) {
-                if (preg_match('/RETS-Session-ID\=(.*?)(\;|\s+|$)/', $cookie, $matches)) {
-                    $this->rets_session_id = $matches[1];
-                }
+            if ($cookie && preg_match('/RETS-Session-ID\=(.*?)(\;|\s+|$)/', (string) $cookie, $matches)) {
+                $this->rets_session_id = $matches[1];
             }
         }
 
         $this->debug('Response: HTTP ' . $response->getStatusCode());
 
-        if (stripos($response->getHeader('Content-Type'), 'text/xml') !== false and $capability != 'GetObject') {
+        if (stripos((string) $response->getHeader('Content-Type'), 'text/xml') !== false && $capability != 'GetObject') {
             $parser = $this->grab(Strategy::PARSER_XML);
             $xml = $parser->parse($response);
 
-            if ($xml and isset($xml['ReplyCode'])) {
-                $rc = (string)$xml['ReplyCode'];
+            if ($xml && isset($xml['ReplyCode'])) {
+                $rc = (string) $xml['ReplyCode'];
 
-                if ($rc == "20037" and $capability != 'Login') {
+                if ($rc == '20037' && $capability != 'Login') {
                     // must make Login request again.  let's handle this automatically
 
                     if ($this->getConfiguration()->readOption('disable_auto_retry')) {
                         // this attempt was already a retry, so let's stop here
                         $this->debug("Re-logging in disabled.  Won't retry");
-                        throw new RETSException($xml['ReplyText'], (int)$xml['ReplyCode']);
+                        throw new RETSException($xml['ReplyText'], (int) $xml['ReplyCode']);
                     }
 
                     if ($is_retry) {
                         // this attempt was already a retry, so let's stop here
                         $this->debug("Request retry failed.  Won't retry again");
-                        // let this error fall through to the more generic handling below
+                    // let this error fall through to the more generic handling below
                     } else {
-                        $this->debug("RETS 20037 re-auth requested");
-                        $this->debug("Logging in again and retrying request");
+                        $this->debug('RETS 20037 re-auth requested');
+                        $this->debug('Logging in again and retrying request');
                         // see if logging in again and retrying the request works
                         $this->Login();
 
@@ -433,17 +446,17 @@ class Session
                 // 20201 - No records found - not exception worthy in my mind
                 // 20403 - No objects found - not exception worthy in my mind
                 if (!in_array($rc, [0, 20201, 20403])) {
-                    throw new RETSException($xml['ReplyText'], (int)$xml['ReplyCode']);
+                    throw new RETSException($xml['ReplyText'], (int) $xml['ReplyCode']);
                 }
             }
         }
 
-        if ($this->getConfiguration()->readOption('getobject_auto_retry') and $capability == 'GetObject') {
-            if (stripos($response->getHeader('Content-Type'), 'multipart') !== false) {
+        if ($this->getConfiguration()->readOption('getobject_auto_retry') && $capability == 'GetObject') {
+            if (stripos((string) $response->getHeader('Content-Type'), 'multipart') !== false) {
                 $parser = $this->grab(Strategy::PARSER_OBJECT_MULTIPLE);
                 $collection = $parser->parse($response);
             } else {
-                $collection = new Collection;
+                $collection = new Collection();
                 $parser = $this->grab(Strategy::PARSER_OBJECT_SINGLE);
                 $object = $parser->parse($response);
                 $collection->push($object);
@@ -451,14 +464,14 @@ class Session
 
             /** @var BaseObject[] $collection */
             foreach ($collection as $object) {
-                if ($object->isError() and $object->getError()->getCode() == '20037') {
+                if ($object->isError() && $object->getError()->getCode() == '20037') {
                     if ($is_retry) {
                         // this attempt was already a retry, so let's stop here
                         $this->debug("Request retry failed.  Won't retry again");
-                        // let this error fall through to the more generic handling below
+                    // let this error fall through to the more generic handling below
                     } else {
-                        $this->debug("RETS 20037 re-auth requested");
-                        $this->debug("Logging in again and retrying request");
+                        $this->debug('RETS 20037 re-auth requested');
+                        $this->debug('Logging in again and retrying request');
                         // see if logging in again and retrying the request works
                         $this->Login();
 
@@ -518,12 +531,12 @@ class Session
     }
 
     /**
-     * @param CookieJarInterface $cookie_jar
      * @return $this
      */
     public function setCookieJar(CookieJarInterface $cookie_jar)
     {
         $this->cookie_jar = $cookie_jar;
+
         return $this;
     }
 
@@ -540,7 +553,7 @@ class Session
      */
     public function getLastResponse()
     {
-        return (string)$this->last_response->getBody();
+        return (string) $this->last_response->getBody();
     }
 
     /**
@@ -551,9 +564,6 @@ class Session
         return $this->client;
     }
 
-    /**
-     * @return mixed
-     */
     public function getRetsSessionId()
     {
         return $this->rets_session_id;
@@ -561,7 +571,6 @@ class Session
 
     /**
      * @param $component
-     * @return mixed
      */
     protected function grab($component)
     {
@@ -577,7 +586,7 @@ class Session
             'auth' => [
                 $this->configuration->getUsername(),
                 $this->configuration->getPassword(),
-                $this->configuration->getHttpAuthenticationMethod()
+                $this->configuration->getHttpAuthenticationMethod(),
             ],
             'headers' => [
                 'User-Agent' => $this->configuration->getUserAgent(),
@@ -585,7 +594,7 @@ class Session
                 'Accept-Encoding' => 'gzip',
                 'Accept' => '*/*',
             ],
-            'curl' => [ CURLOPT_COOKIEFILE => tempnam('/tmp', 'phrets') ]
+            'curl' => [CURLOPT_COOKIEFILE => tempnam('/tmp', 'phrets')],
         ];
 
         // disable following 'Location' header (redirects) automatically
