@@ -1,16 +1,17 @@
-<?php namespace PHRETS\Strategies;
+<?php
+
+namespace PHRETS\Strategies;
 
 use Illuminate\Container\Container;
+use Illuminate\Contracts\Container\BindingResolutionException;
 use PHRETS\Configuration;
 
 class StandardStrategy implements Strategy
 {
     /**
-     * Default components
-     *
-     * @var array
+     * Default components.
      */
-    protected $default_components = [
+    protected array $default_components = [
         Strategy::PARSER_LOGIN => \PHRETS\Parsers\Login\OneFive::class,
         Strategy::PARSER_OBJECT_SINGLE => \PHRETS\Parsers\GetObject\Single::class,
         Strategy::PARSER_OBJECT_MULTIPLE => \PHRETS\Parsers\GetObject\Multiple::class,
@@ -25,42 +26,36 @@ class StandardStrategy implements Strategy
         Strategy::PARSER_XML => \PHRETS\Parsers\XML::class,
     ];
 
-    /**
-     * @var \Illuminate\Container\Container
-     */
-    protected $container;
+    protected Container $container;
 
     /**
      * @param $component
-     * @return mixed
+     *
+     * @throws BindingResolutionException
      */
-    public function provide($component)
+    public function provide($component): mixed
     {
         return $this->container->make($component);
     }
 
     /**
-     * @param Configuration $configuration
      * @return void
      */
     public function initialize(Configuration $configuration)
     {
         // start up the service locator
-        $this->container = new Container;
+        $this->container = new Container();
 
         foreach ($this->default_components as $k => $v) {
-            if ($k == 'parser.login' and $configuration->getRetsVersion()->isAtLeast1_8()) {
+            if ($k == 'parser.login' && $configuration->getRetsVersion()->isAtLeast1_8()) {
                 $v = \PHRETS\Parsers\Login\OneEight::class;
             }
 
-            $this->container->singleton($k, function () use ($v) { return new $v; });
+            $this->container->singleton($k, fn () => new $v());
         }
     }
 
-    /**
-     * @return Container
-     */
-    public function getContainer()
+    public function getContainer(): Container
     {
         return $this->container;
     }
