@@ -1,17 +1,19 @@
-<?php namespace PHRETS\Versions;
+<?php
+
+namespace PHRETS\Versions;
 
 use PHRETS\Exceptions\InvalidRETSVersion;
 
-class RETSVersion
+class RETSVersion implements \Stringable
 {
-    const VERSION_1_5 = '1.5';
-    const VERSION_1_7 = '1.7';
-    const VERSION_1_7_1 = '1.7.1';
-    const VERSION_1_7_2 = '1.7.2';
-    const VERSION_1_8 = '1.8';
+    public const VERSION_1_5 = '1.5';
+    public const VERSION_1_7 = '1.7';
+    public const VERSION_1_7_1 = '1.7.1';
+    public const VERSION_1_7_2 = '1.7.2';
+    public const VERSION_1_8 = '1.8';
 
-    protected $number;
-    protected $valid_versions = [
+    protected string $number;
+    protected array $valid_versions = [
         self::VERSION_1_5,
         self::VERSION_1_7,
         self::VERSION_1_7_1,
@@ -21,119 +23,85 @@ class RETSVersion
 
     /**
      * @param $version
+     *
      * @return $this
+     *
      * @throws InvalidRETSVersion
      */
-    public function setVersion($version)
+    public function setVersion($version): static
     {
-        $this->number = str_replace('RETS/', '', $version);
+        $this->number = str_replace('RETS/', '', (string) $version);
         if (!in_array($this->number, $this->valid_versions)) {
             throw new InvalidRETSVersion("RETS version '{$version}' given is not understood");
         }
+
         return $this;
     }
 
-    /**
-     * @return mixed
-     */
-    public function getVersion()
+    public function getVersion(): string
     {
         return $this->number;
     }
 
-    /**
-     * @return string
-     */
-    public function asHeader()
+    public function asHeader(): string
     {
         return 'RETS/' . $this->number;
     }
 
-    /**
-     * @return bool
-     */
-    public function is1_5()
+    public function is1_5(): bool
     {
-        return ($this->number == self::VERSION_1_5);
+        return $this->number == self::VERSION_1_5;
     }
 
-    /**
-     * @return bool
-     */
-    public function is1_7()
+    public function is1_7(): bool
     {
-        return ($this->number == self::VERSION_1_7);
+        return $this->number == self::VERSION_1_7;
     }
 
-    /**
-     * @return bool
-     */
-    public function is1_7_2()
+    public function is1_7_2(): bool
     {
-        return ($this->number == self::VERSION_1_7_2);
+        return $this->number == self::VERSION_1_7_2;
     }
 
-    /**
-     * @return bool
-     */
-    public function is1_8()
+    public function is1_8(): bool
     {
-        return ($this->number == self::VERSION_1_8);
+        return $this->number == self::VERSION_1_8;
     }
 
     /**
      * @param $version
-     * @return bool
      */
-    public function isAtLeast($version)
+    public function isAtLeast($version): bool
     {
-        return (version_compare($this->number, $version) >= 0);
+        return version_compare($this->number, $version) >= 0;
     }
 
-    /**
-     * @return bool
-     */
-    public function isAtLeast1_5()
+    public function isAtLeast1_5(): bool
     {
         return $this->isAtLeast(self::VERSION_1_5);
     }
 
-    /**
-     * @return bool
-     */
-    public function isAtLeast1_7()
+    public function isAtLeast1_7(): bool
     {
         return $this->isAtLeast(self::VERSION_1_7);
     }
 
-    /**
-     * @return bool
-     */
-    public function isAtLeast1_7_2()
+    public function isAtLeast1_7_2(): bool
     {
         return $this->isAtLeast(self::VERSION_1_7_2);
     }
 
-    /**
-     * @return bool
-     */
-    public function isAtLeast1_8()
+    public function isAtLeast1_8(): bool
     {
         return $this->isAtLeast(self::VERSION_1_8);
     }
-    
-    /**
-     * @return array
-     */
-    public function getValidVersions()
+
+    public function getValidVersions(): array
     {
         return $this->valid_versions;
     }
 
-    /**
-     * @return string
-     */
-    public function __toString()
+    public function __toString(): string
     {
         return $this->asHeader();
     }
