@@ -1,17 +1,19 @@
 <?php
 
-class ClientTest extends PHPUnit_Framework_TestCase {
+use PHPUnit\Framework\TestCase;
 
+class ClientTest extends TestCase
+{
     /** @test **/
-    public function it_makes()
+    public function itMakes()
     {
         $this->assertInstanceOf('GuzzleHttp\\Client', \PHRETS\Http\Client::make());
     }
 
     /** @test **/
-    public function it_allows_overrides()
+    public function itAllowsOverrides()
     {
-        $gc = new GuzzleHttp\Client;
+        $gc = new GuzzleHttp\Client();
         \PHRETS\Http\Client::set($gc);
 
         $this->assertSame($gc, \PHRETS\Http\Client::make());
