@@ -1,9 +1,11 @@
 <?php
 
-class ResponseTest extends PHPUnit_Framework_TestCase {
+use PHPUnit\Framework\TestCase;
 
+class ResponseTest extends TestCase
+{
     /** @test **/
-    public function it_creates_valid_xml()
+    public function itCreatesValidXml()
     {
         $body = "<?xml version='1.0' encoding='UTF-8'?><guestbook><guest><fname>First Name</fname><lname>Last Name</lname></guest></guestbook>";
         $guzzleResponse = new GuzzleHttp\Psr7\Response(200, ['X-Foo' => 'Bar'], $body);
@@ -14,7 +16,7 @@ class ResponseTest extends PHPUnit_Framework_TestCase {
     }
 
     /** @test **/
-    public function it_creates_valid_xml_with_new_lines()
+    public function itCreatesValidXmlWithNewLines()
     {
         $body = "\n\n\r<?xml version='1.0' encoding='UTF-8'?><guestbook><guest><fname>First Name</fname><lname>Last Name</lname></guest></guestbook>\r\n\n";
         $guzzleResponse = new GuzzleHttp\Psr7\Response(200, ['X-Foo' => 'Bar'], $body);
