@@ -1,4 +1,6 @@
-<?php namespace PHRETS;
+<?php
+
+namespace PHRETS;
 
 use PHRETS\Exceptions\InvalidConfiguration;
 use PHRETS\Strategies\StandardStrategy;
@@ -7,8 +9,8 @@ use PHRETS\Versions\RETSVersion;
 
 class Configuration
 {
-    const AUTH_BASIC = 'basic';
-    const AUTH_DIGEST = 'digest';
+    public const AUTH_BASIC = 'basic';
+    public const AUTH_DIGEST = 'digest';
 
     protected $username;
     protected $password;
@@ -24,30 +26,24 @@ class Configuration
 
     public function __construct()
     {
-        $this->rets_version = (new RETSVersion)->setVersion('1.5');
+        $this->rets_version = (new RETSVersion())->setVersion('1.5');
     }
 
-    /**
-     * @return mixed
-     */
     public function getLoginUrl()
     {
         return $this->login_url;
     }
 
     /**
-     * @param mixed $login_url
      * @return $this
      */
     public function setLoginUrl($login_url)
     {
         $this->login_url = $login_url;
+
         return $this;
     }
 
-    /**
-     * @return mixed
-     */
     public function getPassword()
     {
         return $this->password;
@@ -55,11 +51,13 @@ class Configuration
 
     /**
      * @param string $password
+     *
      * @return $this
      */
     public function setPassword($password)
     {
         $this->password = $password;
+
         return $this;
     }
 
@@ -73,17 +71,16 @@ class Configuration
 
     /**
      * @param string $rets_version
+     *
      * @return $this
      */
     public function setRetsVersion($rets_version)
     {
         $this->rets_version = (new RETSVersion())->setVersion($rets_version);
+
         return $this;
     }
 
-    /**
-     * @return mixed
-     */
     public function getUserAgent()
     {
         return $this->user_agent;
@@ -91,17 +88,16 @@ class Configuration
 
     /**
      * @param string $user_agent
+     *
      * @return $this
      */
     public function setUserAgent($user_agent)
     {
         $this->user_agent = $user_agent;
+
         return $this;
     }
 
-    /**
-     * @return mixed
-     */
     public function getUserAgentPassword()
     {
         return $this->user_agent_password;
@@ -109,17 +105,16 @@ class Configuration
 
     /**
      * @param string $user_agent_password
+     *
      * @return $this
      */
     public function setUserAgentPassword($user_agent_password)
     {
         $this->user_agent_password = $user_agent_password;
+
         return $this;
     }
 
-    /**
-     * @return mixed
-     */
     public function getUsername()
     {
         return $this->username;
@@ -127,37 +122,44 @@ class Configuration
 
     /**
      * @param string $username
+     *
      * @return $this
      */
     public function setUsername($username)
     {
         $this->username = $username;
+
         return $this;
     }
 
     /**
      * @param $name
      * @param $value
+     *
      * @return $this
      */
     public function setOption($name, $value)
     {
         $this->options[$name] = $value;
+
         return $this;
     }
 
     /**
      * @param $name
+     *
      * @return null
      */
     public function readOption($name)
     {
-        return (array_key_exists($name, $this->options)) ? $this->options[$name] : null;
+        return $this->options[$name] ?? null;
     }
 
     /**
      * @param array $configuration
+     *
      * @return static
+     *
      * @throws Exceptions\InvalidConfiguration
      */
     public static function load($configuration = [])
@@ -172,7 +174,7 @@ class Configuration
             'http_authentication' => 'HttpAuthenticationMethod',
         ];
 
-        $me = new static;
+        $me = new self();
 
         foreach ($variables as $k => $m) {
             if (array_key_exists($k, $configuration)) {
@@ -182,7 +184,7 @@ class Configuration
         }
 
         if (!$me->valid()) {
-            throw new InvalidConfiguration("Login URL and Username must be provided");
+            throw new InvalidConfiguration('Login URL and Username must be provided');
         }
 
         return $me;
@@ -193,10 +195,7 @@ class Configuration
      */
     public function valid()
     {
-        if (!$this->getLoginUrl() or !$this->getUsername()) {
-            return false;
-        }
-        return true;
+        return $this->getLoginUrl() && $this->getUsername();
     }
 
     /**
@@ -205,37 +204,39 @@ class Configuration
     public function getStrategy()
     {
         if (!$this->strategy) {
-            $this->setStrategy(new StandardStrategy);
+            $this->setStrategy(new StandardStrategy());
         }
+
         return $this->strategy;
     }
 
     /**
-     * @param Strategies\Strategy $strategy
      * @return $this
      */
     public function setStrategy(Strategy $strategy)
     {
         $strategy->initialize($this);
         $this->strategy = $strategy;
+
         return $this;
     }
 
     /**
-     * @param Session $session
      * @return string
      */
     public function userAgentDigestHash(Session $session)
     {
-        $ua_a1 = md5($this->getUserAgent() .':'. $this->getUserAgentPassword());
+        $ua_a1 = md5($this->getUserAgent() . ':' . $this->getUserAgentPassword());
+
         return md5(
-            trim($ua_a1) .'::'. trim($session->getRetsSessionId()) .
-            ':'. trim($this->getRetsVersion()->asHeader())
+            trim((string) $ua_a1) . '::' . trim((string) $session->getRetsSessionId()) .
+            ':' . trim((string) $this->getRetsVersion()->asHeader())
         );
     }
 
     /**
      * @param $auth_method
+     *
      * @return $this
      */
     public function setHttpAuthenticationMethod($auth_method)
@@ -244,6 +245,7 @@ class Configuration
             throw new \InvalidArgumentException("Given authentication method is invalid.  Must be 'basic' or 'digest'");
         }
         $this->http_authentication = $auth_method;
+
         return $this;
     }
 
