@@ -1,11 +1,13 @@
-<?php namespace PHRETS\Interpreters;
+<?php
+
+namespace PHRETS\Interpreters;
 
 class Search
 {
-    public static function dmql($query)
+    public static function dmql(?string $query): ?string
     {
         // automatically surround the given query with parentheses if it doesn't have them already
-        if (!empty($query) and $query != "*" and !preg_match('/^\((.*)\)$/', $query)) {
+        if (!empty($query) && $query != '*' && !preg_match('/^\((.*)\)$/', (string) $query)) {
             $query = '(' . $query . ')';
         }
 
