@@ -1,22 +1,25 @@
-<?php namespace PHRETS\Parsers\GetMetadata;
+<?php
 
-use PHRETS\Http\Response;
+namespace PHRETS\Parsers\GetMetadata;
+
 use Illuminate\Support\Collection;
+use PHRETS\Http\Response;
+use PHRETS\Parsers\XML;
 use PHRETS\Session;
 
 class Resource extends Base
 {
-    public function parse(Session $rets, Response $response)
+    public function parse(Session $rets, Response $response): Collection
     {
-        /** @var \PHRETS\Parsers\XML $parser */
+        /** @var XML $parser */
         $parser = $rets->getConfiguration()->getStrategy()->provide(\PHRETS\Strategies\Strategy::PARSER_XML);
         $xml = $parser->parse($response);
 
-        $collection = new Collection;
+        $collection = new Collection();
 
         if ($xml->METADATA) {
             foreach ($xml->METADATA->{'METADATA-RESOURCE'}->Resource as $key => $value) {
-                $metadata = new \PHRETS\Models\Metadata\Resource;
+                $metadata = new \PHRETS\Models\Metadata\Resource();
                 $metadata->setSession($rets);
                 /** @var \PHRETS\Models\Metadata\Resource $obj */
                 $obj = $this->loadFromXml($metadata, $value, $xml->METADATA->{'METADATA-RESOURCE'});
