@@ -1,29 +1,36 @@
-<?php namespace PHRETS\Parsers\Search;
+<?php
 
-use PHRETS\Http\Response;
+namespace PHRETS\Parsers\Search;
+
 use PHRETS\Exceptions\AutomaticPaginationError;
+use PHRETS\Exceptions\CapabilityUnavailable;
+use PHRETS\Http\Response;
 use PHRETS\Models\Search\Results;
 use PHRETS\Session;
 
 class RecursiveOneX
 {
-    public function parse(Session $rets, Response $response, $parameters)
+    /**
+     * @throws CapabilityUnavailable
+     * @throws AutomaticPaginationError
+     */
+    public function parse(Session $rets, Response $response, $parameters): Results
     {
         // we're given the first response automatically, so parse this and start the recursion
 
-        /** @var \PHRETS\Parsers\Search\OneX $parser */
+        /** @var OneX $parser */
         $parser = $rets->getConfiguration()->getStrategy()->provide('parser.search');
         $rs = $parser->parse($rets, $response, $parameters);
 
         while ($this->continuePaginating($rets, $parameters, $rs)) {
             $pms = $parameters;
 
-            $rets->debug("Continuing pagination...");
-            $rets->debug("Current count collected already: " . $rs->count());
+            $rets->debug('Continuing pagination...');
+            $rets->debug('Current count collected already: ' . $rs->count());
 
             $resource = $pms['SearchType'];
             $class = $pms['Class'];
-            $query = (array_key_exists('Query', $pms)) ? $pms['Query'] : null;
+            $query = $pms['Query'] ?? null;
 
             $pms['Offset'] = $this->getNewOffset($rets, $parameters, $rs);
 
@@ -31,7 +38,6 @@ class RecursiveOneX
             unset($pms['Class']);
             unset($pms['Query']);
 
-            /** @var Results $inner_rs */
             $inner_rs = $rets->Search($resource, $class, $query, $pms, false);
             $rs->setTotalResultsCount($inner_rs->getTotalResultsCount());
             $rs->setMaxRowsReached($inner_rs->isMaxRowsReached());
@@ -50,39 +56,28 @@ class RecursiveOneX
     }
 
     /**
-     * @param Session $rets
      * @param $parameters
-     * @param Results $rs
-     * @return bool
      */
-    protected function continuePaginating(Session $rets, $parameters, Results $rs)
+    protected function continuePaginating(Session $rets, $parameters, Results $rs): bool
     {
         return $rs->isMaxRowsReached();
     }
 
     /**
-     * @param Session $rets
      * @param $parameters
-     * @param Results $rs
-     * @return int
      */
-    protected function getNewOffset(Session $rets, $parameters, Results $rs)
+    protected function getNewOffset(Session $rets, $parameters, Results $rs): int
     {
         return $rs->getReturnedResultsCount() + 1;
     }
 
-    /**
-     * @param Results $big
-     * @param Results $small
-     * @return bool
-     */
-    protected function isPaginationBroken(Results $big, Results $small)
+    protected function isPaginationBroken(Results $big, Results $small): bool
     {
         $big_first = $big->first();
         $small_first = $small->first();
 
-        if ($big_first and $small_first) {
-            return $big_first->toArray() == $small_first->toArray();
+        if ($big_first && $small_first) {
+            return $big_first->toArray() === $small_first->toArray();
         }
 
         return false;
