@@ -1,13 +1,19 @@
-<?php namespace PHRETS\Parsers;
+<?php
 
+namespace PHRETS\Parsers;
+
+use Exception;
 use PHRETS\Http\Response;
 use Psr\Http\Message\ResponseInterface;
 
 class XML
 {
-    public function parse($string)
+    /**
+     * @throws Exception
+     */
+    public function parse($string): \SimpleXMLElement
     {
-        if ($string instanceof ResponseInterface or $string instanceof Response) {
+        if ($string instanceof ResponseInterface || $string instanceof Response) {
             $string = $string->getBody()->__toString();
         }
 
