@@ -1,13 +1,14 @@
 <?php
 
+use PHPUnit\Framework\TestCase;
 use PHRETS\Configuration;
 
-class ConfigurationTest extends PHPUnit_Framework_TestCase {
-
+class ConfigurationTest extends TestCase
+{
     /** @test **/
-    public function it_does_the_basics()
+    public function itDoesTheBasics()
     {
-        $config = new Configuration;
+        $config = new Configuration();
         $config->setLoginUrl('http://www.reso.org/login'); // not a valid RETS server.  just using for testing
         $config->setUsername('user');
         $config->setPassword('pass');
@@ -18,7 +19,7 @@ class ConfigurationTest extends PHPUnit_Framework_TestCase {
     }
 
     /** @test **/
-    public function it_loads_config_from_array()
+    public function itLoadsConfigFromArray()
     {
         $config = Configuration::load([
             'login_url' => 'http://www.reso.org/login',
@@ -33,42 +34,42 @@ class ConfigurationTest extends PHPUnit_Framework_TestCase {
 
     /**
      * @test
-     * @expectedException \PHRETS\Exceptions\InvalidConfiguration
      **/
-    public function it_complains_about_bad_config()
+    public function itComplainsAboutBadConfig()
     {
+        $this->expectException(\PHRETS\Exceptions\InvalidConfiguration::class);
         Configuration::load();
     }
 
     /** @test **/
-    public function it_loads_default_rets_version()
+    public function itLoadsDefaultRetsVersion()
     {
-        $config = new Configuration;
+        $config = new Configuration();
 
         $this->assertInstanceOf('PHRETS\\Versions\\RETSVersion', $config->getRetsVersion());
         $this->assertTrue($config->getRetsVersion()->is1_5());
     }
 
     /** @test **/
-    public function it_handles_versions_correctly()
+    public function itHandlesVersionsCorrectly()
     {
-        $config = new Configuration;
+        $config = new Configuration();
         $config->setRetsVersion('1.7.2');
         $this->assertInstanceOf('PHRETS\\Versions\\RETSVersion', $config->getRetsVersion());
     }
 
     /** @test **/
-    public function it_handles_user_agents()
+    public function itHandlesUserAgents()
     {
-        $config = new Configuration;
+        $config = new Configuration();
         $config->setUserAgent('PHRETS/2.0');
         $this->assertSame('PHRETS/2.0', $config->getUserAgent());
     }
 
     /** @test **/
-    public function it_handles_ua_passwords()
+    public function itHandlesUaPasswords()
     {
-        $config = new Configuration;
+        $config = new Configuration();
         $config->setUserAgent('PHRETS/2.0');
         $config->setUserAgentPassword('test12345');
         $this->assertSame('PHRETS/2.0', $config->getUserAgent());
@@ -76,26 +77,26 @@ class ConfigurationTest extends PHPUnit_Framework_TestCase {
     }
 
     /** @test **/
-    public function it_tracks_options()
+    public function itTracksOptions()
     {
-        $config = new Configuration;
+        $config = new Configuration();
         $config->setOption('param', true);
         $this->assertTrue($config->readOption('param'));
     }
 
     /** @test **/
-    public function it_loads_a_strategy()
+    public function itLoadsAStrategy()
     {
-        $config = new Configuration;
+        $config = new Configuration();
         $this->assertInstanceOf('PHRETS\Strategies\Strategy', $config->getStrategy());
         $this->assertInstanceOf('PHRETS\Strategies\StandardStrategy', $config->getStrategy());
     }
 
     /** @test **/
-    public function it_allows_overriding_the_strategy()
+    public function itAllowsOverridingTheStrategy()
     {
-        $config = new Configuration;
-        $strategy = new \PHRETS\Strategies\StandardStrategy;
+        $config = new Configuration();
+        $strategy = new \PHRETS\Strategies\StandardStrategy();
         $strategy->initialize($config);
         $config->setStrategy($strategy);
 
@@ -103,9 +104,9 @@ class ConfigurationTest extends PHPUnit_Framework_TestCase {
     }
 
     /** @test **/
-    public function it_generates_user_agent_auth_hashes_correctly()
+    public function itGeneratesUserAgentAuthHashesCorrectly()
     {
-        $c = new Configuration;
+        $c = new Configuration();
         $c->setLoginUrl('http://www.reso.org/login')
             ->setUserAgent('PHRETS/2.0')
             ->setUserAgentPassword('12345')
@@ -116,26 +117,26 @@ class ConfigurationTest extends PHPUnit_Framework_TestCase {
     }
 
     /** @test **/
-    public function it_keeps_digest_as_the_default()
+    public function itKeepsDigestAsTheDefault()
     {
-        $c = new Configuration;
+        $c = new Configuration();
         $this->assertSame(Configuration::AUTH_DIGEST, $c->getHttpAuthenticationMethod());
     }
 
     /**
      * @test
-     * @expectedException InvalidArgumentException
      **/
-    public function it_doesnt_allow_bogus_auth_methods()
+    public function itDoesntAllowBogusAuthMethods()
     {
-        $c = new Configuration;
+        $this->expectException(InvalidArgumentException::class);
+        $c = new Configuration();
         $c->setHttpAuthenticationMethod('bogus');
     }
 
     /** @test **/
-    public function it_accepts_basic_auth()
+    public function itAcceptsBasicAuth()
     {
-        $c = new Configuration;
+        $c = new Configuration();
         $c->setHttpAuthenticationMethod(Configuration::AUTH_BASIC);
         $this->assertSame(Configuration::AUTH_BASIC, $c->getHttpAuthenticationMethod());
     }
