@@ -1,48 +1,49 @@
 <?php
 
+use PHPUnit\Framework\TestCase;
 use PHRETS\Interpreters\GetObject;
 
-class GetObjectTest extends PHPUnit_Framework_TestCase {
-
+class GetObjectTest extends TestCase
+{
     /** @test **/
-    public function it_combines_singles()
+    public function itCombinesSingles()
     {
         $this->assertEquals(['12345:1'], GetObject::ids(12345, 1));
     }
 
     /** @test **/
-    public function it_combines_multiple_from_string()
+    public function itCombinesMultipleFromString()
     {
-        $this->assertEquals(['12345:1','67890:1'], GetObject::ids('12345,67890', 1));
+        $this->assertEquals(['12345:1', '67890:1'], GetObject::ids('12345,67890', 1));
     }
 
     /** @test **/
-    public function it_combines_multiple_from_colon_string()
+    public function itCombinesMultipleFromColonString()
     {
-        $this->assertEquals(['12345:1','67890:1'], GetObject::ids('12345:67890', 1));
+        $this->assertEquals(['12345:1', '67890:1'], GetObject::ids('12345:67890', 1));
     }
 
     /** @test **/
-    public function it_combines_multiple_from_array()
+    public function itCombinesMultipleFromArray()
     {
-        $this->assertEquals(['12345:1','67890:1'], GetObject::ids([12345, 67890], 1));
+        $this->assertEquals(['12345:1', '67890:1'], GetObject::ids([12345, 67890], 1));
     }
 
     /** @test **/
-    public function it_combines_multiple_object_id_strings()
+    public function itCombinesMultipleObjectIdStrings()
     {
-        $this->assertEquals(['12345:1:2:3','67890:1:2:3'], GetObject::ids([12345, 67890], '1,2,3'));
+        $this->assertEquals(['12345:1:2:3', '67890:1:2:3'], GetObject::ids([12345, 67890], '1,2,3'));
     }
 
     /** @test **/
-    public function it_combines_multiple_object_id_arrays()
+    public function itCombinesMultipleObjectIdArrays()
     {
-        $this->assertEquals(['12345:1:2:3','67890:1:2:3'], GetObject::ids([12345, 67890], [1, 2, 3]));
+        $this->assertEquals(['12345:1:2:3', '67890:1:2:3'], GetObject::ids([12345, 67890], [1, 2, 3]));
     }
 
     /** @test **/
-    public function it_parses_ranges()
+    public function itParsesRanges()
     {
-        $this->assertEquals(['12345:1:2:3:4:5','67890:1:2:3:4:5'], GetObject::ids([12345, 67890], '1-5'));
+        $this->assertEquals(['12345:1:2:3:4:5', '67890:1:2:3:4:5'], GetObject::ids([12345, 67890], '1-5'));
     }
 }
