@@ -1,23 +1,24 @@
 <?php
 
+use PHPUnit\Framework\TestCase;
 use PHRETS\Interpreters\Search;
 
-class SearchTest extends PHPUnit_Framework_TestCase
+class SearchTest extends TestCase
 {
     /** @test **/
-    public function it_doesnt_touch_properly_formatted_dmql()
+    public function itDoesntTouchProperlyFormattedDmql()
     {
         $this->assertSame('(FIELD=VALUE)', Search::dmql('(FIELD=VALUE)'));
     }
-    
+
     /** @test **/
-    public function it_wraps_simplified_dmql_in_parens()
+    public function itWrapsSimplifiedDmqlInParens()
     {
         $this->assertSame('(FIELD=VALUE)', Search::dmql('FIELD=VALUE'));
     }
 
     /** @test **/
-    public function it_doesnt_modify_when_special_characters_are_used()
+    public function itDoesntModifyWhenSpecialCharactersAreUsed()
     {
         $this->assertSame('*', Search::dmql('*'));
         $this->assertSame('', Search::dmql(''));
