@@ -1,13 +1,14 @@
 <?php
 
+use PHPUnit\Framework\TestCase;
 use PHRETS\Models\Metadata\Resource;
 
-class ResourceTest extends PHPUnit_Framework_TestCase {
-
+class ResourceTest extends TestCase
+{
     /** @test **/
-    public function it_holds()
+    public function itHolds()
     {
-        $metadata = new Resource;
+        $metadata = new Resource();
         $metadata->setDescription('Test Description');
 
         $this->assertSame('Test Description', $metadata->getDescription());
@@ -15,38 +16,38 @@ class ResourceTest extends PHPUnit_Framework_TestCase {
 
     /**
      * @test
-     * @expectedException BadMethodCallException
      **/
-    public function it_doesnt_like_bad_methods()
+    public function itDoesntLikeBadMethods()
     {
-        $metadata = new Resource;
+        $this->expectException(BadMethodCallException::class);
+        $metadata = new Resource();
         $metadata->totallyBogus();
     }
 
     /** @test **/
-    public function it_returns_null_for_unrecognized_attributes()
+    public function itReturnsNullForUnrecognizedAttributes()
     {
-        $metadata = new Resource;
+        $metadata = new Resource();
         $this->assertNull($metadata->getSomethingFake());
     }
-    
+
     /** @test **/
-    public function it_works_like_an_array()
+    public function itWorksLikeAnArray()
     {
-        $metadata = new Resource;
+        $metadata = new Resource();
         $metadata->setDescription('Test Description');
 
         $this->assertTrue(isset($metadata['Description']));
         $this->assertSame('Test Description', $metadata['Description']);
     }
-    
-    /** @test **/
-    public function it_sets_like_an_array()
-    {
-        $metadata = new Resource;
-        $metadata['Description'] = "Array setter";
 
-        $this->assertSame("Array setter", $metadata->getDescription());
+    /** @test **/
+    public function itSetsLikeAnArray()
+    {
+        $metadata = new Resource();
+        $metadata['Description'] = 'Array setter';
+
+        $this->assertSame('Array setter', $metadata->getDescription());
 
         unset($metadata['Description']);
 
