@@ -1,6 +1,6 @@
 <?php
 
-require_once(__DIR__ . "/../vendor/autoload.php");
-require_once(__DIR__ . "/Integration/BaseIntegration.php");
-require_once(__DIR__ . "/Integration/Parsers/CustomSystemParser.php");
-require_once(__DIR__ . "/Integration/Parsers/CustomXMLParser.php");
+require_once __DIR__ . '/../vendor/autoload.php';
+require_once __DIR__ . '/Integration/BaseIntegration.php';
+require_once __DIR__ . '/Integration/Parsers/CustomSystemParser.php';
+require_once __DIR__ . '/Integration/Parsers/CustomXMLParser.php';
